Servie

+++ src/main/java/com/demo/service/ConsumerService.java
@@ -1,6 +1,9 @@
 package com.demo.service;
 
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jdbc.repository.query.Modifying;
 import org.springframework.stereotype.Service;
 import com.demo.repo.ConsumerRepository;
 
@@ -9,13 +12,19 @@ public class ConsumerService {
 	@Autowired
 	ConsumerRepository consumerRepository;
 	
-	public void updateByConsumerId(String name, String area,String city, String connectionType,int id) {
+	public boolean updateByConsumerId(String name, String area,String city, String connectionType,int id) {
 		if(consumerRepository.existsById(id)) {
 			consumerRepository.updateByConsumerId(name, area, city, connectionType, id);
+			return true;
 		}
+		return false;	
 	}
 	
-	public void deleteByConsumerId(int id) {
-		consumerRepository.deleteById(id);
+	public boolean deleteByConsumerId(int id) {
+		if(consumerRepository.existsById(id)) {
+			consumerRepository.deleteById(id);
+			return true;
+		}
+		return false;
 	}
 }
